+++ src/Symfony/Component/Validator/Constraints/Count.php
@@ -11,7 +11,9 @@
 
 namespace Symfony\Component\Validator\Constraints;
 
+use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 use Symfony\Component\Validator\Constraint;
+use Symfony\Component\Validator\Exception\LogicException;
 use Symfony\Component\Validator\Exception\MissingOptionsException;
 
 /**
@@ -40,6 +42,7 @@ class Count extends Constraint
     public $min;
     public $max;
     public $divisibleBy;
+    public $condition;
 
     /**
      * {@inheritdoc}
@@ -57,11 +60,16 @@ class Count extends Constraint
         string $divisibleByMessage = null,
         array $groups = null,
         $payload = null,
-        array $options = []
+        array $options = [],
+        string $condition = null
     ) {
         if (\is_array($exactly)) {
             $options = array_merge($exactly, $options);
             $exactly = $options['value'] ?? null;
+        }
+
+        if (null !== $condition && !class_exists(ExpressionLanguage::class)) {
+            throw new LogicException(sprintf('The "symfony/expression-language" component is required to use the "%s" constraint with a condition expression.', __CLASS__));
         }
 
         $min = $min ?? $options['min'] ?? null;
@@ -82,6 +90,7 @@ class Count extends Constraint
         $this->minMessage = $minMessage ?? $this->minMessage;
         $this->maxMessage = $maxMessage ?? $this->maxMessage;
         $this->divisibleByMessage = $divisibleByMessage ?? $this->divisibleByMessage;
+        $this->condition = $condition ?? $this->condition;
 
         if (null === $this->min && null === $this->max && null === $this->divisibleBy) {
             throw new MissingOptionsException(sprintf('Either option "min", "max" or "divisibleBy" must be given for constraint "%s".', __CLASS__), ['min', 'max', 'divisibleBy']);